requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-generativeai>=0.8.5
selenium>=4.38.0
webdriver-manager>=4.0.2
